Remove model_fields_set check and rename files to drop underscore prefix
- Removed Pydantic model_fields_set runtime detection from annotation lookup
  (no longer needed since pydantic is not a dependency)
- Renamed _annotation_lookup.py -> annotation_lookup.py
- Renamed _generic_resolution.py -> generic_resolution.py
- Removed tests that depended on model_fields_set behavior
